Change turn to uint64.
This led to changing scent values to float64

+++ example/muncher/util.go
@@ -63,8 +63,8 @@ func timeMe(start time.Time, name string) {
 
 }
 
-func distance(x0 int, y0 int, x1 int, y1 int) float32 {
+func distance(x0 int, y0 int, x1 int, y1 int) float64 {
 	x := x1 - x0
 	y := y1 - y0
-	return float32(math.Sqrt(float64(x*x) + float64(y*y)))
+	return math.Sqrt(float64(x*x) + float64(y*y))
 }
